// src/modules/insights/domain/insight-flag.enum.ts
// src/modules/insights/domain/insight-flag.enum.ts

export enum InsightFlag {
  LOW_CTR = 'LOW_CTR',
  HIGH_BOUNCE_RATE = 'HIGH_BOUNCE_RATE',
  PERF_LCP_REGRESSION = 'PERF_LCP_REGRESSION',
  PERF_CLS_REGRESSION = 'PERF_CLS_REGRESSION',
  PERF_INP_REGRESSION = 'PERF_INP_REGRESSION',
  PERF_TTFB_REGRESSION = 'PERF_TTFB_REGRESSION',
  ENGAGEMENT_DISTRIBUTION_DROP = 'ENGAGEMENT_DISTRIBUTION_DROP',
  ENGAGEMENT_TREND_DOWN = 'ENGAGEMENT_TREND_DOWN',
  VISIBILITY_DROP = 'VISIBILITY_DROP',
  ANOMALY_DETECTED = 'ANOMALY_DETECTED',
  NEW_BEST_PERFORMER = 'NEW_BEST_PERFORMER',
  FUNNEL_BOTTLENECK = 'FUNNEL_BOTTLENECK',
  PERFORMANCE_REGRESSION = 'PERFORMANCE_REGRESSION',
  COHORT_IMBALANCE = 'COHORT_IMBALANCE',
  HIGH_FORM_ABANDON = 'HIGH_FORM_ABANDON',
  LOW_FORM_COMPLETION = 'LOW_FORM_COMPLETION',
  HIGH_FORM_ERRORS = 'HIGH_FORM_ERRORS',
  HIGH_RAGE_CLICKS = 'HIGH_RAGE_CLICKS',
  HIGH_EXIT_INTENT = 'HIGH_EXIT_INTENT',
}

export enum InsightSeverity {
  HIGH = 'high',
  MEDIUM = 'medium',
  LOW = 'low',
  INFO = 'info',
}